IS-8 Support for selecting authentication provider
Core OpenID Connect does not offer mechanisms for specifying an authentication authority/provider in the same was SAML does.

This PR introduces a request parameter that a Relying Party can use to specify the desired authentication provider to OP:s that support several authentication mechanisms.

Such OP:s are either proxies or simply a service that offers several different authentication methods for the user to choose from.

The primary use case is that the RP can use this feature when re-authenticating a user and the RP wants to ensure that the user doesn't have to select the authentication method once again.