IS-82 Signing requirements
In order to have a better way of detecting which key the sender used, the use of `kid` in JOSE headers are set to SHOULD (also in cases where the entity's JWK set only contains one signing key). The reason is mainly to enable dynamic re-load of JWKS-documents after key rollover.